feat: update app routing and all link elements
