--- src/components/forms/index.js
@@ -1,1 +1,0 @@
-export { default as Login } from "./Login";
