+++ src/components/pages/index.js
@@ -1,3 +1,3 @@
 export { default as Home } from "./Home";
 export { default as About } from "./About";
-export { default as Auth } from "./Auth";
+export { default as Login } from "./Login";
